requirements.txt
paho-mqtt==1.5.1
pyserial
